requisicoes.py foi renomada para funcoes_aux.py. requisicoes.py passa a guardar apenas as requisições principais para atualizar os cards no pipefy
